// apps/webapp/src/interfaces/AdminTableRow.ts
interface AdminTableRow {
  id: number;
  name?: string;
  selectedCourse?: string;
  entryDate?: string;
  lastActivity?: string;
  studentStatus?: string;
  enrolledCourse?: string;
  courseGroup?: string;
  bill?: string;
  paymentDate?: string;
  description?: string;
  paymentMethod?: string;
  total?: string;
  groupId?: string;
  course?: string;
  career?: string;
  semester?: string;
  enrolled?: string;
  available?: string;
  status?: string;
}

export default AdminTableRow;